move vector operations to vector_processing
